Harden 1.21.11 runtime checks

+++ src/main/java/net/werdei/configloader/ConfigLoader.java
@@ -6,6 +6,7 @@ import net.fabricmc.loader.api.FabricLoader;
 import java.io.File;
 import java.io.FileReader;
 import java.io.FileWriter;
+import java.io.IOException;
 
 public abstract class ConfigLoader
 {
@@ -27,11 +28,19 @@ public abstract class ConfigLoader
     public static void save(Class<?> configClass, String fileName) throws ConfigException
     {
         File file = new File(FabricLoader.getInstance().getConfigDir().toFile(), fileName);
+        File parent = file.getParentFile();
+        if (parent != null && !parent.exists() && !parent.mkdirs())
+            throw new ConfigException("Error creating config directory " + parent, null);
+
         try (FileWriter writer = new FileWriter(file))
         {
             writer.write(gsonBuilder.create().toJson(configClass.getConstructor().newInstance()));
         }
-        catch (Exception e)
+        catch (IOException e)
+        {
+            throw new ConfigException("Error writing configuration file " + file, e);
+        }
+        catch (ReflectiveOperationException e)
         {
             throw new ConfigException("Error creating a temporary instance of class " + configClass.getName(), e);
         }
